Add semantic versioning operations
Semantic versioning is important and can be used in PDA data storage
or in a program to provide different execution paths for different
program versions or data versions.

The `Semver` type introduced here will allow these operations to happen
without the need for external parsers and will be a compact array if stored onchain.

- A `Utils` struct has been added  with support for parsing a `u8` into an ASCII character
- A `HullSvmResult` type has been added as a helper for `Result<T, HullSvmError>
- A `HullSvmError` type has been added to handle errors
- `ordering` feature is used to enable PartialOrd and Ord for sorting
- `display_and_debug` is used to enable Display and Debug features

These code changes are in the `hull-svm-common` crate

Closes #3