Add the builtin system accounts
Builtin program are important for SVM to functions as intented.
This commit add all the programs from `solana-sdk-ids` crate in a manner useful for
this crate. It loads the address, the loader of the program, the short name of the
program as defined in the `solana_sdk_ids::...` modules and parses the base58 public key
into bytes

This add a chore to exlude the account information from typos cli since
addresses are not guaranteed to be valid words